Adjust Jest config
